Simple synq order update endpoint test
Includes renaming of restapi module to hostapi module to semantically
separate API for host messages from API for SynQ messages and others as
they come.

This includes an attempt to separate SecurityConfig and SwaggerConfig for
these two APIs as well, however I need input if this can be done better.
I would prefer to have configs for SynQ api in the synqapi module.