MLT-0038 Callback for Orders (#39)
Simple synq order update endpoint implementation

Includes renaming of restapi module to hostapi module to semantically
separate API for host messages from API for SynQ messages and others as
they come. Making sure that the rest of the naming makes sense the 
product was changed to item, except for SynQ module which uses the word
product.

This includes an attempt to separate SecurityConfig and SwaggerConfig for
these two APIs as well. This works for now but might be problematic if we
use other ways of authentication for SynQ or other systems