MLT-0072 Emails for Orders (#58)
* Create a basic mail sender

* Rework Mail system to use Spring Mail

- Fix missing dependency for WLSService Bean
- Also fixed tests related to this
- Remove MailConfig as it can be purely handled via Spring
- Add Testcontainer for e-mail via Mailhog

The last point is important. It allows for future handling
of e-mail related tests. It is also kinda required as
creating a Mock Bean of the JavaMailSender
when an application context is present requires
all the Spring-related dependencies.

E.G. MailHealthContributor, which brings in the MailHealthIndicator etc.

* Add EmailRepository, with a Mongo implementation

* Restructuring

* Split order and mail integration test

* Update + disable broken email test

* Create mail structure for storage handlers

* Fix bulk find method in ItemMongoRepository

Uses proper Spring Data Mongo syntax, which fixes
a suppressed exception which this method was throwing

* Use FreeMarker for HTML-based emails

* Prototype for making order confirmation PDFs

* Generate and inline QR codes in emails

* Remove PDF handling

Not worth to implement at the moment.
Might be after cleaning up the EmailAdapter, but not right now

* Some restructuring and cleanup

* Remove unused fields in email model

* Add license header for QR image generation

* Swap to Kotlin require

* Structure cleanup and smaller QR size for order-id

* Address reviews

- Add version property for qrcodegen
- Change QR-code colors
- Use different e-mail address

* Move email handling to separate function

* Update email structure

- Uses description lists for fields
- Translate order type
- Unify sending logic

* Add QR string to alt text

* Rename environment and config variables

Warehouse references changed to order handler instead

* Translate order type and make titles bold in email templates

Specifically translates the storages' order type

* Add a sorting TODO in pom

* Rename order email creation methods

* Add mock email for test

* Allow order confirmations to be disabled

* Update K8s deployments and README to include email server env

* Allow for email server to be disabled

This allows for full disabling the email adapter by simply not
providing the env-variables Spring needs
to create a JavaMailSender.

* Use environment condition instead of bean condition

Makes it so that the test environment will properly initialize
the JavaMailSender bean.