Introduce UpdateItem
Splits some functionality off from MoveItem, which has
been changed to now only modify the quantity of the item.
This is different from the original, where it used to override it.

UpdateItem preserves the previous behavior of hard overriding it.
All of this is required to correctly handle messages from AutoStore.