MLT-0141 Add UpdateItem (#94)
* Introduce UpdateItem

Splits some functionality off from MoveItem, which has
been changed to now only modify the quantity of the item.
This is different from the original, where it used to override it.

UpdateItem preserves the previous behavior of hard overriding it.
All of this is required to correctly handle messages from AutoStore.

* Update tests

* Use better exception for quantity error

* John Spotless is truly the nemesis of programming

* Improve test readability

* Add schema for quantity on hand

* Add test for moving products

* Fix payloads being in the incorrect order

* Change quantityMove validation to require value above one

* Simply using var for quantity

Co-authored-by: Daniel Aaron Salwerowicz <Daniel@Salwerowi.cz>

* Bring balance to the nine realms by removing a space

---------

Co-authored-by: Daniel Aaron Salwerowicz <Daniel@Salwerowi.cz>